using const_iterators using intializer lists explicitly using quint64 type instead of unsigned long long, it is both shorter to type and more expresive in that it is a specific bit width
svn path=/trunk/KDE/kdeutils/kcalc/; revision=1168870

+++ kcalc_bitset.cpp
@@ -110,7 +110,7 @@ KCalcBitset::KCalcBitset(QWidget *parent)
     }
 }
 
-void KCalcBitset::setValue(unsigned long long value)
+void KCalcBitset::setValue(quint64 value)
 {
     if (value_ == value) return;
 
@@ -122,14 +122,14 @@ void KCalcBitset::setValue(unsigned long long value)
     }
 }
 
-unsigned long long KCalcBitset::getValue()
+quint64 KCalcBitset::getValue()
 {
     return value_;
 }
 
 void KCalcBitset::slotToggleBit(int bit)
 {
-    unsigned long long nv = getValue() ^(1LL << bit);
+    quint64 nv = getValue() ^(1LL << bit);
     setValue(nv);
     emit valueChanged(value_);
 }
